Added Cart Surfer
I've now managed to have it so that if you go to the min shack and click on the doorway you can play cart surfer! It is a one time thing currently as you play through the game and I haven't coded it to send you anywhere yet. Also the inventory and code etc don't disappear when you play so it looks kinda weird lol